+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `bar` component tokens (Orange-OpenSource/ouds-ios#1169)
 - `effect` raw and semantic tokens (Orange-OpenSource/ouds-ios#1169)
+- Read only variant for `checkbox` and  `checkbox indeterminate` component (Orange-OpenSource/ouds-ios#1137)
 - Icon flip option for `text input` component trailing action (Orange-OpenSource/ouds-ios#1041)
 - Icon flip option for `button` component (Orange-OpenSource/ouds-ios#1124)
 - Support for `watchOS` (Orange-OpenSource/ouds-ios#1110)